feat: inapp api 제작

+++ services/service-inapp/src/main/kotlin/com/b1nd/dodamdodam/inapp/application/team/TeamUseCase.kt
@@ -8,6 +8,8 @@ import com.b1nd.dodamdodam.inapp.application.team.data.request.CreateTeamRequest
 import com.b1nd.dodamdodam.inapp.application.team.data.request.DeleteTeamMemberRequest
 import com.b1nd.dodamdodam.inapp.application.team.data.request.DeleteTeamRequest
 import com.b1nd.dodamdodam.inapp.application.team.data.request.EditTeamInfoRequest
+import com.b1nd.dodamdodam.inapp.application.team.data.response.MyTeamResponse
+import com.b1nd.dodamdodam.inapp.application.team.data.toMyTeamResponses
 import com.b1nd.dodamdodam.inapp.domain.team.service.TeamService
 import org.springframework.stereotype.Component
 import org.springframework.transaction.annotation.Transactional
@@ -23,6 +25,12 @@ class TeamUseCase(
         return Response.created("팀이 생성되었어요.")
     }
 
+    fun getMyTeam(): Response<List<MyTeamResponse>> {
+        val userId = PassportHolder.current().requireUserId()
+        val teamMemberList = teamService.getAllByUser(userId);
+        return Response.ok("내 팀들을 조회했어요.", teamMemberList.toMyTeamResponses())
+    }
+
     fun addTeamMembers(request: AddTeamMemberRequest): Response<Any> {
         val team = teamService.getById(request.teamId)
         teamService.addMember(team, request.users)
@@ -30,16 +38,19 @@ class TeamUseCase(
     }
 
     fun deleteTeamMembers(request: DeleteTeamMemberRequest): Response<Any> {
+        teamService.validateOwner(PassportHolder.current().requireUserId(), request.teamId)
         teamService.deleteMember(request.teamId, request.users)
         return Response.ok("팀 멤버가 삭제되었어요.")
     }
 
     fun deleteTeam(request: DeleteTeamRequest): Response<Any> {
+        teamService.validateOwner(PassportHolder.current().requireUserId(), request.teamId)
         teamService.delete(request.teamId)
         return Response.ok("팀이 삭제되었어요.")
     }
 
     fun editTeamInfo(request: EditTeamInfoRequest): Response<Any> {
+        teamService.validateOwner(PassportHolder.current().requireUserId(), request.teamId)
         teamService.updateInfo(
             request.teamId,
             request.name,
